+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- [#48](https://github.com/alchemauss/mauss/pull/48): add `unique` utility function
 - [#47](https://github.com/alchemauss/mauss/pull/47): remove default exports except `api` module
 - [#44](https://github.com/alchemauss/mauss/pull/44): enable `checkJs` and `resolveJsonModule`